# Conflicts:
#	package.json
#	yarn.lock

+++ src/HPApi.ts
@@ -1,26 +1,31 @@
 "use strict";
 
 import WalkupScanDestination, {
-  WalkupScanDestinationData
+  WalkupScanDestinationData,
 } from "./WalkupScanDestination";
-
+import WalkupScanToCompDestination, {
+  WalkupScanToCompDestinationData,
+} from "./WalkupScanToCompDestination";
 import util from "util";
 import fs from "fs";
 import axios, { AxiosResponse } from "axios";
-import url from "url";
+import { URL } from "url";
 import xml2js from "xml2js";
 import EventTable, { EventTableData } from "./EventTable";
 import Job, { JobData } from "./Job";
 import ScanStatus, { ScanStatusData } from "./ScanStatus";
 import WalkupScanDestinations, {
-  WalkupScanDestinationsData
+  WalkupScanDestinationsData,
 } from "./WalkupScanDestinations";
+import WalkupScanToCompDestinations, {
+  WalkupScanToCompDestinationsData,
+} from "./WalkupScanToCompDestinations";
 import ScanJobSettings from "./ScanJobSettings";
 import Destination from "./Destination";
 import { Stream } from "stream";
 
 const parser = new xml2js.Parser();
-const parseString = util.promisify(parser.parseString);
+const parseString = util.promisify<string, any>(parser.parseString);
 let printerIP = "192.168.1.11";
 
 export default class HPApi {
@@ -33,7 +38,7 @@ export default class HPApi {
       baseURL: `http://${printerIP}`,
       url: "/WalkupScan/WalkupScanDestinations",
       method: "GET",
-      responseType: "text"
+      responseType: "text",
     });
 
     if (response.status !== 200) {
@@ -46,14 +51,50 @@ export default class HPApi {
     }
   }
 
+  static async getWalkupScanToCompDestinations(): Promise<WalkupScanToCompDestinations> {
+    const response = await axios({
+      baseURL: `http://${printerIP}`,
+      url: "/WalkupScanToComp/WalkupScanToCompDestinations",
+      method: "GET",
+      responseType: "text",
+    });
+
+    if (response.status !== 200) {
+      throw new Error(response.statusText);
+    } else {
+      const parsed = (await parseString(
+        response.data
+      )) as WalkupScanToCompDestinationsData;
+      return new WalkupScanToCompDestinations(parsed);
+    }
+  }
+
+  static async getWalkupScanToCompCaps(): Promise<boolean> {
+    return await axios({
+      baseURL: `http://${printerIP}`,
+      url: "/WalkupScanToComp/WalkupScanToCompCaps",
+      method: "GET",
+      responseType: "text",
+    }).then(
+      (response) => response.status == 200,
+      () => false
+    );
+  }
+
   static async removeDestination(walkupScanDestination: WalkupScanDestination) {
-    let urlInfo = url.parse(walkupScanDestination.resourceURI);
+    let urlInfo = new URL(walkupScanDestination.resourceURI);
+
+    if (urlInfo.pathname === null) {
+      throw new Error(
+        `invalid walkupScanDestination.resourceURI: ${walkupScanDestination.resourceURI}`
+      );
+    }
 
     const response = await axios({
       baseURL: `http://${printerIP}`,
       url: urlInfo.pathname,
       method: "DELETE",
-      responseType: "text"
+      responseType: "text",
     });
     if (response.status === 204) {
       return true;
@@ -62,19 +103,23 @@ export default class HPApi {
     }
   }
 
-  static async registerDestination(destination: Destination) {
-    const xml = await destination.toXML();
+  static async registerDestination(destination: Destination, toComp: boolean) {
+    let xml = await destination.toXML();
+    let url = "/WalkupScan/WalkupScanDestinations";
+    if (toComp) {
+      url = "/WalkupScanToComp/WalkupScanToCompDestinations";
+    }
     const response = await axios({
       baseURL: `http://${printerIP}`,
-      url: "/WalkupScan/WalkupScanDestinations",
+      url: url,
       method: "POST",
       headers: { "Content-Type": "text/xml" },
       data: xml,
-      responseType: "text"
+      responseType: "text",
     });
 
     if (response.status === 201) {
-      return response.headers.location;
+      return new URL(response.headers.location).pathname;
     } else {
       throw response;
     }
@@ -84,7 +129,7 @@ export default class HPApi {
     etag = "",
     timeout = 0
   ): Promise<{ etag: string; eventTable: EventTable }> {
-    let url = this.appendTimeout(timeout, "/EventMgmt/EventTable");
+    let url = this.appendTimeout("/EventMgmt/EventTable", timeout);
 
     let headers = this.placeETagHeader(etag, {});
 
@@ -95,14 +140,14 @@ export default class HPApi {
         url: url,
         method: "GET",
         responseType: "text",
-        headers: headers
+        headers: headers,
       });
     } catch (error) {
       response = error.response;
       if (response.status === 304) {
         return {
           etag: etag,
-          eventTable: new EventTable({})
+          eventTable: new EventTable({}),
         };
       }
       throw error;
@@ -111,20 +156,20 @@ export default class HPApi {
     const parsed = await parseString(response.data);
     return {
       etag: response.headers["etag"],
-      eventTable: new EventTable(parsed as EventTableData)
+      eventTable: new EventTable(parsed as EventTableData),
     };
   }
 
   static placeETagHeader(etag: string, headers: object) {
     if (etag !== "") {
       headers = {
-        "If-None-Match": etag
+        "If-None-Match": etag,
       };
     }
     return headers;
   }
 
-  static appendTimeout(timeout: number | null = null, url: string): string {
+  static appendTimeout(url: string, timeout: number | null = null): string {
     if (timeout == null) {
       timeout = 1200;
     }
@@ -136,18 +181,26 @@ export default class HPApi {
 
   static async getDestination(destinationURL: string) {
     const response = await axios({
+      baseURL: `http://${printerIP}`,
       url: destinationURL,
       method: "GET",
-      responseType: "text"
+      responseType: "text",
     });
 
     if (response.status !== 200) {
       throw response;
     } else {
-      const parsed = (await parseString(
-        response.data
-      )) as WalkupScanDestinationData;
-      return new WalkupScanDestination(parsed);
+      if (destinationURL.includes("WalkupScanToComp")) {
+        const parsed = (await parseString(
+          response.data
+        )) as WalkupScanToCompDestinationData;
+        return new WalkupScanToCompDestination(parsed);
+      } else {
+        const parsed = (await parseString(
+          response.data
+        )) as WalkupScanDestinationData;
+        return new WalkupScanDestination(parsed);
+      }
     }
   }
 
@@ -156,7 +209,7 @@ export default class HPApi {
       baseURL: `http://${printerIP}`,
       url: "/Scan/Status",
       method: "GET",
-      responseType: "text"
+      responseType: "text",
     });
 
     if (response.status !== 200) {
@@ -168,7 +221,7 @@ export default class HPApi {
   }
 
   static delay(t: number) {
-    return new Promise(function(resolve) {
+    return new Promise(function (resolve) {
       setTimeout(resolve, t);
     });
   }
@@ -182,7 +235,7 @@ export default class HPApi {
       method: "POST",
       headers: { "Content-Type": "text/xml" },
       data: xml,
-      responseType: "text"
+      responseType: "text",
     });
 
     if (response.status === 201) {
@@ -200,7 +253,7 @@ export default class HPApi {
     const response = await axios({
       url: jobURL,
       method: "GET",
-      responseType: "text"
+      responseType: "text",
     });
 
     if (response.status !== 200) {
@@ -219,7 +272,7 @@ export default class HPApi {
       baseURL: `http://${printerIP}:8080`,
       url: binaryURL,
       method: "GET",
-      responseType: "stream"
+      responseType: "stream",
     });
 
     const destinationFileStream = fs.createWriteStream(destination);
